+++ .flake8
@@ -17,5 +17,7 @@ exclude =
     # __init__s usually have pre-imported modules
     model/**/__init__.py
     data/**/__init__.py
+    optimizers/__init__.py
+    optimizers/radam.py
 max_complexity = 10
 statistics = true
