+++ tests/test_utils.py
@@ -3,35 +3,9 @@ from unittest import TestCase
 import numpy as np
 
 from utils import (
-    hard_max,
     get_2d_from_3d,
     get_3d_from_2d,
 )
-
-
-class MetricTestCase(TestCase):
-
-    def setUp(self):
-        self.x = np.array(
-            [[
-                [0.1, 0.5, 0.7],
-                [0.9, 0.5, 0.3],
-            ]]
-        )
-
-    def test_hard_max(self):
-        hard_max_x = hard_max(self.x)
-        self.assertEqual(
-            np.all(
-                hard_max_x == np.array(
-                    [[
-                        [0., 1., 1.],
-                        [1., 0., 0.],
-                    ]]
-                )
-            ),
-            True,
-        )
 
 
 class DimensionConvertTestCase(TestCase):
